backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
pandas
pyarrow
tables   # Para HDF5
numpy
python-multipart